Separated group task routes. Separated shared task functions
