+++ website/server/controllers/api-v3/tasks.js
@@ -9,72 +9,16 @@ import {
   NotAuthorized,
   BadRequest,
 } from '../../libs/errors';
+import {
+  _createTasks,
+  _getTasks,
+} from '../../libs/api-v3/taskManager';
 import common from '../../../../common';
 import Bluebird from 'bluebird';
 import _ from 'lodash';
 import logger from '../../libs/logger';
 
 let api = {};
-
-async function _validateTaskAlias (tasks, res) {
-  let tasksWithAliases = tasks.filter(task => task.alias);
-  let aliases = tasksWithAliases.map(task => task.alias);
-
-  // Compares the short names in tasks against
-  // a Set, where values cannot repeat. If the
-  // lengths are different, some name was duplicated
-  if (aliases.length !== [...new Set(aliases)].length) {
-    throw new BadRequest(res.t('taskAliasAlreadyUsed'));
-  }
-
-  await Bluebird.map(tasksWithAliases, (task) => {
-    return task.validate();
-  });
-}
-
-// challenge must be passed only when a challenge task is being created
-async function _createTasks (req, res, user, challenge, group) {
-  let toSave = Array.isArray(req.body) ? req.body : [req.body];
-
-  toSave = toSave.map(taskData => {
-    // Validate that task.type is valid
-    if (!taskData || Tasks.tasksTypes.indexOf(taskData.type) === -1) throw new BadRequest(res.t('invalidTaskType'));
-
-    let taskType = taskData.type;
-    let newTask = new Tasks[taskType](Tasks.Task.sanitize(taskData));
-
-    if (challenge) {
-      newTask.challenge.id = challenge.id;
-    } else if (group) {
-      newTask.group.id = group.id;
-    } else {
-      newTask.userId = user._id;
-    }
-
-    // Validate that the task is valid and throw if it isn't
-    // otherwise since we're saving user/challenge and task in parallel it could save the user/challenge with a tasksOrder that doens't match reality
-    let validationErrors = newTask.validateSync();
-    if (validationErrors) throw validationErrors;
-
-    // Otherwise update the user/challenge
-    if (!group) (challenge || user).tasksOrder[`${taskType}s`].unshift(newTask._id);
-
-    return newTask;
-  });
-
-  // tasks with aliases need to be validated asyncronously
-  await _validateTaskAlias(toSave, res);
-
-  toSave = toSave.map(task => task.save({ // If all tasks are valid (this is why it's not in the previous .map()), save everything, withough running validation again
-    validateBeforeSave: false,
-  }));
-
-  toSave.unshift((challenge || user).save());
-
-  let tasks = await Bluebird.all(toSave);
-  tasks.splice(0, 1); // Remove user or challenge
-  return tasks;
-}
 
 /**
  * @api {post} /api/v3/tasks/user Create a new task belonging to the user
@@ -90,7 +34,8 @@ api.createUserTasks = {
   url: '/tasks/user',
   middlewares: [authWithHeaders()],
   async handler (req, res) {
-    let tasks = await _createTasks(req, res, res.locals.user);
+    let user = res.locals.user;
+    let tasks = await _createTasks(req, res, {user});
     res.respond(201, tasks.length === 1 ? tasks[0] : tasks);
   },
 };
@@ -125,7 +70,7 @@ api.createChallengeTasks = {
     if (!challenge || user.challenges.indexOf(challengeId) === -1) throw new NotFound(res.t('challengeNotFound'));
     if (challenge.leader !== user._id) throw new NotAuthorized(res.t('onlyChalLeaderEditTasks'));
 
-    let tasks = await _createTasks(req, res, user, challenge);
+    let tasks = await _createTasks(req, res, {user, challenge});
 
     res.respond(201, tasks.length === 1 ? tasks[0] : tasks);
 
@@ -135,108 +80,6 @@ api.createChallengeTasks = {
     return null;
   },
 };
-
-/**
- * @api {post} /api/v3/tasks/group/:groupId Create a new task belonging to a group
- * @apiDescription Can be passed an object to create a single task or an array of objects to create multiple tasks.
- * @apiVersion 3.0.0
- * @apiName CreateGroupTasks
- * @apiGroup Task
- *
- * @apiParam {UUID} groupId The id of the group the new task(s) will belong to
- *
- * @apiSuccess data An object if a single task was created, otherwise an array of tasks
- */
-api.createGroupTasks = {
-  method: 'POST',
-  url: '/tasks/group/:groupId',
-  middlewares: [authWithHeaders()],
-  async handler (req, res) {
-    // req.checkParams('groupId', res.t('groupIdRequired')).notEmpty().isUUID();
-
-    let reqValidationErrors = req.validationErrors();
-    if (reqValidationErrors) throw reqValidationErrors;
-
-    let user = res.locals.user;
-
-    let group = await Group.getGroup({user, groupId: req.params.groupId, populateLeader: false});
-    if (!group) throw new NotFound(res.t('groupNotFound'));
-
-    if (group.leader !== user._id) throw new NotAuthorized(res.t('onlyGroupLeaderEditTasks'));
-
-    let tasks = await _createTasks(req, res, user, null, group);
-
-    res.respond(201, tasks.length === 1 ? tasks[0] : tasks);
-
-    return null;
-  },
-};
-
-// challenge must be passed only when a challenge task is being created
-// @TODO: Factory - GetTasksQuery(type) - where type is user, challenge, group
-async function _getTasks (req, res, user, challenge, group) {
-  let query = {userId: user._id};
-
-  if (challenge) {
-    query =  {'challenge.id': challenge.id, userId: {$exists: false}};
-  } else if (group) {
-    query =  {'group.id': group.id, userId: {$exists: false}};
-  }
-
-  let type = req.query.type;
-
-  if (type) {
-    if (type === 'todos') {
-      query.completed = false; // Exclude completed todos
-      query.type = 'todo';
-    } else if (type === 'completedTodos' || type === '_allCompletedTodos') { // _allCompletedTodos is currently in BETA and is likely to be removed in future
-      let limit = 30;
-
-      if (type === '_allCompletedTodos') {
-        limit = 0; // no limit
-      }
-      query = Tasks.Task.find({
-        userId: user._id,
-        type: 'todo',
-        completed: true,
-      }).limit(limit).sort({
-        dateCompleted: -1,
-      });
-    } else {
-      query.type = type.slice(0, -1); // removing the final "s"
-    }
-  } else {
-    query.$or = [ // Exclude completed todos
-      {type: 'todo', completed: false},
-      {type: {$in: ['habit', 'daily', 'reward']}},
-    ];
-  }
-
-  let tasks = await Tasks.Task.find(query).exec();
-
-  // Order tasks based on tasksOrder
-  if (type && type !== 'completedTodos' && type !== '_allCompletedTodos') {
-    let order = (challenge || user).tasksOrder[type];
-    let orderedTasks = new Array(tasks.length);
-    let unorderedTasks = []; // what we want to add later
-
-    tasks.forEach((task, index) => {
-      let taskId = task._id;
-      let i = order[index] === taskId ? index : order.indexOf(taskId);
-      if (i === -1) {
-        unorderedTasks.push(task);
-      } else {
-        orderedTasks[i] = task;
-      }
-    });
-
-    // Remove empty values from the array and add any unordered task
-    orderedTasks = _.compact(orderedTasks).concat(unorderedTasks);
-    res.respond(200, orderedTasks);
-  } else {
-    res.respond(200, tasks);
-  }
-}
 
 /**
  * @api {get} /api/v3/tasks/user Get a user's tasks
@@ -260,7 +103,9 @@ api.getUserTasks = {
     let validationErrors = req.validationErrors();
     if (validationErrors) throw validationErrors;
 
-    return await _getTasks(req, res, res.locals.user);
+    let user = res.locals.user;
+
+    return await _getTasks(req, res, {user});
   },
 };
 
@@ -295,39 +140,7 @@ api.getChallengeTasks = {
     let group = await Group.getGroup({user, groupId: challenge.group, fields: '_id type privacy', optionalMembership: true});
     if (!group || !challenge.canView(user, group)) throw new NotFound(res.t('challengeNotFound'));
 
-    return await _getTasks(req, res, res.locals.user, challenge);
-  },
-};
-
-/**
- * @api {get} /api/v3/tasks/group/:groupId Get a group's tasks
- * @apiVersion 3.0.0
- * @apiName GetGroupTasks
- * @apiGroup Task
- *
- * @apiParam {UUID} groupId The id of the group from which to retrieve the tasks
- * @apiParam {string="habits","dailys","todos","rewards"} type Optional query parameter to return just a type of tasks
- *
- * @apiSuccess {Array} data An array of tasks
- */
-api.getGroupTasks = {
-  method: 'GET',
-  url: '/tasks/group/:groupId',
-  middlewares: [authWithHeaders()],
-  async handler (req, res) {
-    req.checkParams('groupId', res.t('groupIdRequired')).notEmpty().isUUID();
-    let types = Tasks.tasksTypes.map(type => `${type}s`);
-    req.checkQuery('type', res.t('invalidTaskType')).optional().isIn(types);
-
-    let validationErrors = req.validationErrors();
-    if (validationErrors) throw validationErrors;
-
-    let user = res.locals.user;
-
-    let group = await Group.getGroup({user, groupId: req.params.groupId, populateLeader: false});
-    if (!group) throw new NotFound(res.t('groupNotFound'));
-
-    return await _getTasks(req, res, res.locals.user, null, group);
+    return await _getTasks(req, res, {user, challenge});
   },
 };
